requirements.txt
numpy>=1.17.4
Pillow>=6.2.1
PyGLM>=1.1.2
bottle>=0.12.17
paste>=3.2.3
urlpath>=1.1.4
